changed some class names to avoid collision with other plugins.
